core review refactoring

+++ admin/name_update.php
@@ -2,21 +2,40 @@
 
 include(__DIR__ . '/../include/include.php');
 
-$num_get = (int) f_igosja_request_get('num');
+if (!$num_get = (int) f_igosja_request_get('num'))
+{
+    redirect('/wrong_page.php');
+}
+
+$sql = "SELECT `name_id`,
+               `name_name`
+        FROM `name`
+        WHERE `name_id`=$num_get
+        LIMIT 1";
+$name_sql = f_igosja_mysqli_query($sql, false);
+
+if (0 == $name_sql->num_rows)
+{
+    redirect('/wrong_page.php');
+}
+
+$name_array = $name_sql->fetch_all(1);
 
 if ($data = f_igosja_request_post('data'))
 {
-    $set_sql = f_igosja_sql_data($data);
+    $set_sql = f_igosja_sql_data($data, array(
+        'name_name'
+    ));
 
     $sql = "UPDATE `name`
             SET $set_sql
             WHERE `name_id`=$num_get
             LIMIT 1";
-    f_igosja_mysqli_query($sql);
+    f_igosja_mysqli_query($sql, false);
 
     $sql = "DELETE FROM `namecountry`
             WHERE `namecountry_name_id`=$num_get";
-    f_igosja_mysqli_query($sql);
+    f_igosja_mysqli_query($sql, false);
 
     $country = f_igosja_request_post('array', 'namecountry_country_id');
 
@@ -27,37 +46,23 @@ if ($data = f_igosja_request_post('data'))
         $sql = "INSERT INTO `namecountry`
                 SET `namecountry_name_id`=$num_get,
                     `namecountry_country_id`=$country_id";
-        f_igosja_mysqli_query($sql);
+        f_igosja_mysqli_query($sql, false);
     }
 
     redirect('/admin/name_view.php?num=' . $num_get);
 }
 
-$sql = "SELECT `name_id`,
-               `name_name`
-        FROM `name`
-        WHERE `name_id`=$num_get
-        LIMIT 1";
-$name_sql = f_igosja_mysqli_query($sql);
-
-if (0 == $name_sql->num_rows)
-{
-    redirect('/wrong_page.php');
-}
-
-$name_array = $name_sql->fetch_all(1);
-
 $sql = "SELECT `namecountry_country_id`
         FROM `namecountry`
         WHERE `namecountry_name_id`=$num_get";
-$namecountry_sql = f_igosja_mysqli_query($sql);
+$namecountry_sql = f_igosja_mysqli_query($sql, false);
 $namecountry_array = $namecountry_sql->fetch_all(1);
 
 $sql = "SELECT `country_id`,
                `country_name`
         FROM `country`
-        ORDER BY `country_name` ASC";
-$country_sql = f_igosja_mysqli_query($sql);
+        ORDER BY `country_name` ASC, `country_id` ASC";
+$country_sql = f_igosja_mysqli_query($sql, false);
 
 $country_array = $country_sql->fetch_all(1);
